🎨 refactor Commands.ts
