--- src/classes/Commands/export.ts
@@ -1,10 +1,0 @@
-import * as botStatus from './statusCommands';
-import * as help from './helpCommands';
-import messageCommand from './messageCommand';
-import * as misc from './miscCommands';
-import * as pricelist from './pricelistManagerCommands';
-import * as review from './reviewCommands';
-import * as utils from './utils';
-import * as options from './optionsCommands';
-
-export { botStatus, help, messageCommand, misc, pricelist, review, utils, options };
