+++ src/classes/Commands/functions/options.ts
@@ -5,14 +5,14 @@ import SteamID from 'steamid';
 import { promises as fsp } from 'fs';
 import * as path from 'path';
 
-import Bot from '../Bot';
-import CommandParser from '../CommandParser';
-import { JsonOptions, removeCliOptions } from '../Options';
+import Bot from '../../Bot';
+import CommandParser from '../../CommandParser';
+import { JsonOptions, removeCliOptions } from '../../Options';
 
-import { deepMerge } from '../../lib/tools/deep-merge';
-import validator from '../../lib/validator';
-import log from '../../lib/logger';
-import MyHandler from '../MyHandler/MyHandler';
+import { deepMerge } from '../../../lib/tools/deep-merge';
+import validator from '../../../lib/validator';
+import log from '../../../lib/logger';
+import MyHandler from '../../MyHandler/MyHandler';
 
 export function optionsCommand(steamID: SteamID, bot: Bot): void {
     const liveOptions = deepMerge({}, bot.options) as JsonOptions;
